requirements.txt
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.1.5
Flask-WTF==0.14.2
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-slugify==3.0.3
SQLAlchemy==1.3.7
Werkzeug==0.16.0
WTForms==2.2.1
-e .
